// clients/theplot/src/app/stories/model/avatar-shape.ts
export type AvatarShape =
  | 'Circle'
  | 'Square'
  | 'Squircle'
  | 'Teardrop'
  | 'Pill'
  | 'Diamond'
  | 'Triangle'
  | 'Pentagon'
  | 'Hexagon'
  | 'Octagon'
  | 'Star'
  | 'Shield';

export const AVATAR_SHAPES: readonly AvatarShape[] = [
  'Circle',
  'Square',
  'Squircle',
  'Teardrop',
  'Pill',
  'Diamond',
  'Triangle',
  'Pentagon',
  'Hexagon',
  'Octagon',
  'Star',
  'Shield',
];